Move files to separe rows
